✨feat(source-google-drive): Introduce File Based Stream Permissions Reader (#55689)
